GRW-400 Use object instead of function + add feature flag + folder restructuring
